pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "vpn-rating-watcher"
version = "0.1.0"
description = "Service for monitoring VPN ratings and publishing daily Telegram updates."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "vpn-rating-watcher contributors" }
]
dependencies = [
  "aiogram>=3.14.0,<4.0.0",
  "alembic>=1.14.0,<2.0.0",
  "matplotlib>=3.9.0,<4.0.0",
  "playwright>=1.48.0,<2.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "pydantic-settings>=2.6.0,<3.0.0",
  "pydantic>=2.9.0,<3.0.0",
  "python-dateutil>=2.9.0,<3.0.0",
  "sqlalchemy>=2.0.36,<3.0.0",
  "typer>=0.12.5,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "pytest-cov>=5.0.0,<6.0.0",
  "ruff>=0.8.0,<1.0.0",
]

[project.scripts]
vrw = "vpn_rating_watcher.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/vpn_rating_watcher"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "SIM"]
ignore = ["UP017"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
